Replace monotonicity rule for vsetvl; make vl CSR read-only
Ensure that the first stripmine iteration of reduction loops uses the
largest vector length of all iterations, even in the case of (AVL < 2*MAXVL).
This allows software to avoid needing to explicitly calculate a running
maximum of vector lengths.  (This situation was the justification for
the original monotonicity requirement across the entire domain of
positive integers.)