Disallow rd = v0 for vadc/vsbc
These instructions implicitly write the carry-out into v0.  So if rd = v0,
there is a WAW hazard within the instruction itself.  Rather than making the
behavior well defined, which could be annoying to implement for some
microarchitectures but provides little benefit to software, I propose we make
the instructions illegal in this case.