Unify vl=0 and vstart >= vl rules
Currently, vl=0 is a special case in the spec: elements past vl are left
intact in this case.  vstart >= vl (when vl > 0) is sort of a special case
itself: no elements below vl are updated, but elements past vl are zeroed.

The proposal is to unfiy these cases, and leave the entire vector intact
when vstart >= vl.  Or, put another way, the elements past vl are zeroed if
and only if some other element in the vector needs to be written.