Remove the vmford instruction
@jhauser-us's survey of the C math library functions concluded that vmford
will not have sufficient impact on performance to justify including.  Many
of the quiet comparisons are against a constant, in which case it does not
help.  For the remaining ones, the penalty is only 1-2 instructions,
depending on whether signaling NaNs are supported.  While signaling NaNs
are supported within the C library itself, GCC's default is to not support
them, so compiled code incurs only a one-instruction penalty for lacking
the vmford instruction.