Add overlap constraint for segment load destination and mask register
If an exception occurs partway through loading a segment, and the mask
register has been overwritten, restarting the load is problematic.

It seems we already came to this conclusion once, beacuse this constraint
exists for indexed segment loads.  The constraint seems to apply equally
to the other classes of segment load, though.