clarify masked-off vs inactive elements
new definitions:
inactive elements are those elements not updated in an instruction invocation (prestart + masked-off + tail)
active elements are those elements updated (body && selected by a mask)

the inactive element set is now larger, as it includes both tail and prestart elements. the rest of spec was updated to be more precise about the term 'inactive'. clarify that prestart elements are always undisturbed.