Clarify that whole-register loads do use the encoded EEW
The old text points out that, because the in-memory layout and in-register
layouts are the same, the instructions behave the same as moving vectors
of bytes with EEW=8.  True enough as far as the data transfer, but not so
for the handling of evl/vstart: those actually do depend on the encoded EEW.

This choice is consistent with the later statement that "implementations are
allowed to raise a misaligned address exception on whole register loads",
meaning (a) that EEW is not truly a HINT and (b) the unit of transfer really
is elements, not bytes, even if the effect of a successful transfer is the
same.