"Clarify" that vlm/vsm respect vtype.vill
The vlm/vsm instructions are only meant to be used when vill=0, since
they are designed to spill or fill only the active subset of a mask
register, rather than a whole register.

Recently, the argument was made that they don't truly depend on vtype,
since their behavior does not depend on the vma, vta, vsew, or vlmul
fields.  However, the instructions are obviously useless when vill=1,
since vill=1 implies vl=0, and so the instructions would become NOPs.

My take is that the instructions indirectly depend on vtype, since
they depend on vl, and vtype governs the set of values vl may assume.
It's not clear that my reasoning is ironclad, but I do believe it to
be in line with what the spec authors intended.